# backend/openapi/openapi.v1.json
{
  "info": {
    "title": "MoneyTracker.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
    "/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "MoneyTracker.Api"
        ]
      }
    },
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "MoneyTracker.Api"
        ]
      }
    },
    "/households/invitations/{invitationToken}/accept": {
      "post": {
        "description": "Adds the authenticated user as a household member.",
        "operationId": "AcceptHouseholdInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "invitationToken",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Accept a household invitation.",
        "tags": [
          "CreateHouseholdEndpoint"
        ]
      }
    },
    "/households/{householdId}/invite": {
      "post": {
        "description": "Only the household owner may issue an invitation token.",
        "operationId": "InviteHouseholdMember",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Invite a user to a household.",
        "tags": [
          "CreateHouseholdEndpoint"
        ]
      }
    },
    "/households/{householdId}/members": {
      "get": {
        "description": "Returns all members for an existing household.",
        "operationId": "GetHouseholdMembers",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Get household members.",
        "tags": [
          "CreateHouseholdEndpoint"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "MoneyTracker.Api"
    },
    {
      "name": "CreateHouseholdEndpoint"
    }
  ]
}